### 2025-03-01 #7862 @arfon
@Romain-Thomas-Shef – thanks for your submission to JOSS. A couple of things:

- Could you please merge your main/default branch into your paper branch so that the code checks (lines of code, license check etc.) work?
- Please rename your paper to `paper.md`

Cheers
Arfon
