+++ paper.md
@@ -1,0 +1,70 @@
+---
+title: 'STON: SofTware for petrOgraphic visualisatioN'
+tags:
+  - Python
+  - Archaeology
+  - Image exploration
+  - Petrography
+authors:
+  - name: Romain Thomas
+    orcid: 0000-0001-8385-3276
+    equal-contrib: true
+    affiliation: "1" # (Multiple affiliations must be quoted)
+  - name: Evgenia Dammer
+    orcid: 0000-0003-0870-9324
+    equal-contrib: true # (This is how you can denote equal contributions between multiple authors)
+    affiliation: 2
+affiliations:
+ - name: RSE team, The University of Sheffield, Regent Court, The University of Sheffield, 211 Portobello St, Sheffield S1 4DP 
+   index: 1
+ - name: Rathgen Research Laboratory, Staatliche Museen zu Berlin, Stiftung Preussischer Kulturbesitz, Schloßstrasse 1A, 14059 Berlin
+   index: 2
+date: 24 February 2025
+bibliography: joss_paper.bib
+
+---
+
+# Summary
+
+`STON` (SofTware for petrOgraphic visualisatioN) is a tool designed to simplify the inspection of microscope images. Originally developed for petrography, specifically for analyzing mineral images, it can be used with any type of image. STON allows users to closely examine images, adjust parameters (contrast, brightness, sharpness), compare images side by side, combine multiple images and analyse features.
+Built with user-friendliness in mind, STON is fully customizable. It is developed in Python using the [PySide6](https://wiki.qt.io/Qt_for_Python) library for the GUI, Pillow [@Murray:2025] and matplotlib [@Hunter:2007] for image display and operation, and scikit-image [@vanderwalt:2014] for deeper image processing. 
+
+# Statement of need
+
+Thin-section petrography is a widely used technique in archaeology for analyzing the composition of ceramic and stone objects, as well as investigating their production technology and provenance [@Reedy:2008; @Peterson:2009; @Quinn:2022]. This method involves studying these materials in thin sections mounted on glass slides under a polarizing microscope to examine their microscopic features. A critical aspect of archaeological study is the comparison and identification of patterns within these features across multiple samples. However, it is typically only possible to view one sample at a time under the microscope. As a result, this method relies heavily on visual memory and repeated observations, making the process inefficient and time-consuming, particularly when dealing with hundreds of samples.
+
+This tool is designed to address these challenges by enabling users to observe multiple photomicrographs simultaneously within a single, convenient interface. It facilitates detailed comparisons, clustering, and data recording, which is especially important in ceramic paste analysis. By allowing users to view multiple samples side by side, the software supports efficient sample grouping and evaluation of compositional characteristics.
+
+While tailored for the specific requirements of petrographic analysis in archaeology, the software has broader applications in other research fields that rely on visual image analysis. By providing an efficient and scalable tool for comparative analysis, it enhances research processes across various disciplines.
+
+# The graphical user interface
+
+STON is a graphical user interface consisting of a main window (see \autoref{fig:GUI}) and multiple sub-windows for performing specific tasks. The main window serves as a central platform where images are loaded from disk and displayed as thumbnails.
+
+![Main window of STON.\label{fig:GUI}](figures_paper/all_GUI_corrected.png)
+
+
+Designed to streamline researchers' work, STON provides a comprehensive suite of tools for image analysis and visual inspection.
+It offers a user-friendly interface and visualization features, STON simplifies the process of inspecting, comparing, and manipulating images.
+All these tools are accessible from the main window, enabling users to explore images in greater detail, extract meaningful insights, and conduct in-depth analyses with ease.
+
[…]
