+++ paper.bib
@@ -1,0 +1,105 @@
+@Article{Hunter:2007,
+  Author    = {Hunter, J. D.},
+  Title     = {Matplotlib: A 2D graphics environment},
+  Journal   = {Computing in Science \& Engineering},
+  Volume    = {9},
+  Number    = {3},
+  Pages     = {90--95},
+  abstract  = {Matplotlib is a 2D graphics package used for Python for
+  application development, interactive scripting, and publication-quality
+  image generation across user interfaces and operating systems.},
+  publisher = {IEEE COMPUTER SOC},
+  doi       = {10.1109/MCSE.2007.55},
+  year      = 2007
+}
+
+@software{Murray:2025,
+  author       = {Andrew Murray and
+                  Hugo van Kemenade and
+                  wiredfool and
+                  Jeffrey A. Clark and
+                  Alexander Karpinsky and
+                  Ondrej Baranovič and
+                  Christoph Gohlke and
+                  Yay295 and
+                  Jon Dufresne and
+                  Matthew Brett and
+                  DWesl and
+                  David Schmidt and
+                  Konstantin Kopachev and
+                  Alastair Houghton and
+                  REDxEYE and
+                  Sandro Mani and
+                  Steve Landey and
+                  Aarni Koskela and
+                  Russell Keith-Magee and
+                  Josh Ware and
+                  vashek and
+                  Piolie and
+                  Stanislau T. and
+                  Jason Douglas and
+                  Uriel Martinez and
+                  David Caro and
+                  Steve Kossouho},
+  title        = {python-pillow/Pillow: 11.1.0},
+  month        = jan,
+  year         = 2025,
+  publisher    = {Zenodo},
+  version      = {11.1.0},
+  doi          = {10.5281/zenodo.14586647},
+  url          = {https://doi.org/10.5281/zenodo.14586647},
+  swhid        = {swh:1:dir:efc42d9bd5cb91aa1da48ebf78f865951cc24c92
+                   ;origin=https://doi.org/10.5281/zenodo.596518;visi
+                   t=swh:1:snp:2108258a783770248cf8b62a080edba61a0c29
+                   b0;anchor=swh:1:rel:6330d5126e83f0864231970e96d0ea
+                   8d99289a97;path=python-pillow-Pillow-4c1aed8
+                  },
+}
+
+@book{55b41ec2-3fba-3289-8b88-65bc008c9228,
+ ISBN = {9781931534550},
+ URL = {http://www.jstor.org/stable/j.ctt3fgvbq},
+ abstract = {As part of the INSTAP Archaeological Excavation Manual series, Thin-Section Petrography of Ceramic Materials provides a concise overview of the history and application of the practice while detailing how this type of petrographic analysis can benefit archaeologists in the field. When thin-section analysis is employed as part of a thorough, multi-disciplinary study of ceramic materials, it provides a wealth of additional interpretative data to archaeologists, allowing for more accurate interpretations of the past, especially regarding pottery production, provenance, variations in technology over time and space, exchange networks on local and non-local scales, and even social issues such as choices of both manufacturers and consumers and traditions of manufacture.},
+ author = {Sarah E. Peterson and Philip P. Betancourt},
+ publisher = {INSTAP Academic Press},
+ title = {Thin-Section Petrography of Ceramic Materials: INSTAP Archaeological Excavation Manual 2},
+ urldate = {2025-02-12},
+ year = {2009}
+}
+
+
+@book{Reedy:2008,
[…]
