+++ paper.md
@@ -20,7 +20,7 @@ affiliations:
  - name: Rathgen Research Laboratory, Staatliche Museen zu Berlin, Stiftung Preussischer Kulturbesitz, Schloßstrasse 1A, 14059 Berlin
    index: 2
 date: 24 February 2025
-bibliography: joss_paper.bib
+bibliography: paper.bib
 
 ---
 
