### 2025-03-02 fe9b4f2
move figures_paper dir in doc
