### 2025-03-02 #7862 @Romain-Thomas-Shef
Hi @arfon!
Thanks for picking up the submission!
I have moved the paper to the main branch and rename it `paper.md`. Hopefully now everything works fine.
