### 2025-04-03 #7862 @arfon
@steko, @koenedaele, @Chilipp, @sgrieve – :wave: would any of you be willing to review this submission for JOSS? The submission under consideration is the Python package STON: Software for petrographic visualisation

Many thanks
Arfon
