### 2025-04-15 #7862 @arfon
@kwinkunks @whimian @Yannan-Chen @jonatanalvelid – 👋 would any of you be willing to review this submission for JOSS? The submission under consideration is the Python package STON: Software for petrographic visualisation

Many thanks
Arfon
