### 2025-04-29 #7862 @arfon
@peterhys @emilikv @kasasxav @cr458 – 👋 would any of you be willing to review this submission for JOSS? The submission under consideration is the Python package STON: Software for petrographic visualisation

Many thanks
Arfon
