### 2025-05-01 #8144 @editorialbot
[REVIEW]: STON: Software for petrographic visualisation
**Submitting author:** @Romain-Thomas-Shef (<a href="http://orcid.org/0000-0001-8385-3276">Romain Thomas</a>)
**Repository:** https://github.com/Romain-Thomas-Shef/STON
**Branch with paper.md** (empty if default branch): main
**Version:** 25.2.1
