### 2025-06-06 #8144 @peterhys
@Romain-Thomas-Shef Hi, I have several comments and questions so far:

1. In the article proof above, line 18, one of the references is not found.
2. In the documentation, I would have a line saying something like "to start the software use command ston in the command line", this might not be so obvious for some people.
3. Can you comment on how this is different from image analysis tools like ImageJ? ImageJ can also do the analysis and can view multiple images at the same time? I would also cite some of the existing tools.
