### 2025-06-10 Romain-Thomas-Shef/STON#22 @Chilipp
add statement of need to documentation
hey here from the JOSS-review https://github.com/openjournals/joss-reviews/issues/8144#issuecomment-2871180699.

>  **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is?

I am missing the statement of need in the documentation at https://romain-thomas-shef.github.io/STON/. Could you please add one? Maybe just the one from the README or the paper?
