### 2025-06-10 Romain-Thomas-Shef/STON#23 @Chilipp
add example data
hey here from the JOSS-review https://github.com/openjournals/joss-reviews/issues/8144#issuecomment-2950436910

> **Example usage:** Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

there is a getting started section, but one is missing the central element to get started, some test data. Could you please add some test data and add a section in *Getting started* how to use this test data?
