### 2025-07-20 Romain-Thomas-Shef/STON#23 @Romain-Thomas-Shef
Hi @Chilipp,

Thanks for opening the issue. Some test data are already in there. You can find how to test the data if you look into the '[testing STON](https://romain-thomas-shef.github.io/STON/tests.html)' section of the documentation.
