### 2025-07-20 #8144 @Romain-Thomas-Shef
Replied to both issues from @Chilipp. @steko You said you wanted to open a few issues. Is it still planned?
