### 2025-07-24 #8144 @Romain-Thomas-Shef
The unprocessed citation is now rendered properly.
