### 2025-08-28 Romain-Thomas-Shef/STON#23 @Romain-Thomas-Shef
Hi @Chilipp .

Did you have time to review the previous message?
Thanks!
