### 2025-08-28 Romain-Thomas-Shef/STON#22 @Romain-Thomas-Shef
@Chilipp,
Could you have a look at the new version of the documentation?
