+++ paper.md
@@ -33,9 +33,11 @@ Built with user-friendliness in mind, STON is fully customizable. It is develope
 
 Thin-section petrography is a widely used technique in archaeology for analyzing the composition of ceramic and stone objects, as well as investigating their production technology and provenance [@Reedy:2008; @Peterson:2009; @Quinn:2022]. This method involves studying these materials in thin sections mounted on glass slides under a polarizing microscope to examine their microscopic features. A critical aspect of archaeological study is the comparison and identification of patterns within these features across multiple samples. However, it is typically only possible to view one sample at a time under the microscope. As a result, this method relies heavily on visual memory and repeated observations, making the process inefficient and time-consuming, particularly when dealing with hundreds of samples.
 
-This tool is designed to address these challenges by enabling users to observe multiple photomicrographs simultaneously within a single, convenient interface. It facilitates detailed comparisons, clustering, and data recording, which is especially important in ceramic paste analysis. By allowing users to view multiple samples side by side, the software supports efficient sample grouping and evaluation of compositional characteristics.
+STON is designed to address these challenges by enabling users to observe multiple photomicrographs simultaneously within a single, convenient interface. It facilitates detailed comparisons, clustering, and data recording, which is especially important in ceramic paste analysis. By allowing users to view multiple samples side by side, the software supports efficient sample grouping and evaluation of compositional characteristics.
 
-While tailored for the specific requirements of petrographic analysis in archaeology, the software has broader applications in other research fields that rely on visual image analysis. By providing an efficient and scalable tool for comparative analysis, it enhances research processes across various disciplines.
+Many existing image analysis tools used in microscopy (e. g. JMicrovision, ImageJ) provide basic functions such as image acquisition, single-image measurement, and annotation. However, these tools often focus on individual images rather than supporting the systematic comparison of large image sets, which is crucial for comprehensive petrographic and archaeometric research. They typically require manual handling of each image and often lack features for integrating multiple images into cohesive overviews, which are essential for detecting patterns and making reliable comparisons across samples, thereby supporting the archaeological interpretation process. Additionally, many of these tools include numerous plugins and complex functions that are not always straightforward or user-friendly for petrographic analysis, making the workflow more cumbersome for researchers. Commercial imaging software (e.g. JENOPTIK, PETROG), frequently tied to subscription models or bundled with hardware, is often limited by licensing costs and restricted flexibility, impeding open collaboration and reproducibility.
+
+STON addresses these limitations by providing an open-source, user-friendly platform specifically designed for petrographic image comparison. It offers tools for efficient image sorting, automated measurements, and the creation of composite overviews, enabling streamlined analysis of large datasets without requiring advanced technical skills or costly licenses. While tailored for the specific requirements of petrographic analysis in archaeology, the software has broader applications in other research fields that rely on visual image analysis. By providing an efficient and scalable tool for comparative analysis, it enhances research processes across various disciplines.
 
 # The graphical user interface
 
