### 2025-09-22 Romain-Thomas-Shef/STON#22 @Romain-Thomas-Shef
@Chilipp would you have time to have a look so we can move forward?
