### 2025-09-30 a01fc38
reviewer comment implementation
