### 2025-09-30 Romain-Thomas-Shef/STON#23 @Romain-Thomas-Shef
Hi @Chilipp,

Thank you for the suggestion. I have added a slightly modified version in the first section of the [getting started](https://romain-thomas-shef.github.io/STON/get_started.html) page.:

> In this part of the documentation you will learn how to start STON and configure it. If you are interested about the GUI please have a look at the [Main window](https://github.com/Romain-Thomas-Shef/STON/issues/main_window.html) page. If you want to explore how the GUI works but do not yet have data ready to work with, you can use the test data that is distributed with STON (see [Testing STON](https://github.com/Romain-Thomas-Shef/STON/issues/tests.html)).
