### 2025-09-30 cd3c0d9
update paper ref
