+++ paper.bib
@@ -56,7 +56,7 @@
                   },
 }
 
-@book{55b41ec2-3fba-3289-8b88-65bc008c9228,
+@book{Peterson:2009,
  ISBN = {9781931534550},
  URL = {http://www.jstor.org/stable/j.ctt3fgvbq},
  abstract = {As part of the INSTAP Archaeological Excavation Manual series, Thin-Section Petrography of Ceramic Materials provides a concise overview of the history and application of the practice while detailing how this type of petrographic analysis can benefit archaeologists in the field. When thin-section analysis is employed as part of a thorough, multi-disciplinary study of ceramic materials, it provides a wealth of additional interpretative data to archaeologists, allowing for more accurate interpretations of the past, especially regarding pottery production, provenance, variations in technology over time and space, exchange networks on local and non-local scales, and even social issues such as choices of both manufacturers and consumers and traditions of manufacture.},
