### 2025-10-22 Romain-Thomas-Shef/STON#23 @Romain-Thomas-Shef
@Chilipp, could you have a look at the changes?
